+++ local_jon/test_001.R
@@ -33,14 +33,14 @@ Carmel_Valley %>%
 # ----- CO ---------------------------------------------------------------------
 
 Salinas_CO <-
-  epa_aqs_loadAnnual(2016, 42101, archiveBaseDir = "~/Data/monitoring") %>%
+  epa_aqs_loadAnnual(42101, 2016, archiveBaseDir = "~/Data/monitoring") %>%
   monitor_filterDate(20160801, 20160810, timezone = "America/Los_Angeles") %>%
   monitor_filterMeta(deviceDeploymentID == "c9fe9c3c096b858a_060531003_01")
 
 monitor_timeseriesPlot(Salinas_CO, shadedNight = TRUE)
 
 Los_FLores_Canyon <-
-  epa_aqs_loadAnnual(2016, 42101, archiveBaseDir = "~/Data/monitoring") %>%
+  epa_aqs_loadAnnual(42101, 2016, archiveBaseDir = "~/Data/monitoring") %>%
   monitor_filterDate(20160614, 20160623, timezone = "America/Los_Angeles") %>%
   monitor_filterMeta(deviceDeploymentID == "8e3f33bcad43c1e2_060831025_01")
 
@@ -49,14 +49,14 @@ monitor_timeseriesPlot(Los_FLores_Canyon, shadedNight = TRUE, col = 'black', opa
 # ----- OZONE ------------------------------------------------------------------
 
 CA_Ozone <-
-  epa_aqs_loadAnnual(2016, 44201, archiveBaseDir = "~/Data/monitoring") %>%
+  epa_aqs_loadAnnual(44201, 2016, archiveBaseDir = "~/Data/monitoring") %>%
   monitor_filterDate(20160614, 20160623, timezone = "America/Los_Angeles") %>%
   monitor_filterMeta(stateCode == "CA")
 
 monitor_leaflet(CA_Ozone)
 
 Upland <-
-  epa_aqs_loadAnnual(2016, 44201, archiveBaseDir = "~/Data/monitoring") %>%
+  epa_aqs_loadAnnual(44201, 2016, archiveBaseDir = "~/Data/monitoring") %>%
   monitor_select("1bec9c5ac76d6d07_060711004_02")
 
 monitor_timeseriesPlot(Upland, addAQI = FALSE)
@@ -66,14 +66,14 @@ addAQILines(pollutant = "OZONE")
 # ----- PM10 -------------------------------------------------------------------
 
 CA_PM10 <-
-  epa_aqs_loadAnnual(2016, 81102, archiveBaseDir = "~/Data/monitoring") %>%
+  epa_aqs_loadAnnual(81102, 2016, archiveBaseDir = "~/Data/monitoring") %>%
   monitor_filterDate(20160614, 20160623, timezone = "America/Los_Angeles") %>%
   monitor_filterMeta(stateCode == "CA")
 
 monitor_leaflet(CA_PM10)
 
 Los_FLores_Canyon <-
-  epa_aqs_loadAnnual(2016, 81102, archiveBaseDir = "~/Data/monitoring") %>%
+  epa_aqs_loadAnnual(81102, 2016, archiveBaseDir = "~/Data/monitoring") %>%
   monitor_filterDate(20160614, 20160623, timezone = "America/Los_Angeles") %>%
   monitor_select("8e3f33bcad43c1e2_060831025_03")